Have done service layer without Vote.

+++ src/main/java/com/restaurant_vote/to/RestaurantWithVotesCountTo.java
@@ -6,9 +6,9 @@ import java.math.BigInteger;
 public class RestaurantWithVotesCountTo extends BaseNamedTo{
     private Integer count;
 
-    public RestaurantWithVotesCountTo(Integer id,String name, BigInteger count) {
+    public RestaurantWithVotesCountTo(Integer id,String name, int count) {
         super(id,name);
-        this.count = count.intValue();
+        this.count = count;
     }
 
     public Integer getCount() {
